Introduce data processing architecture
The architecture consists of a FactExtractor class that contains the
logic required to extract facts from data. The processing takes place
based off of a start date and an end date. The processing logic must be
idempotent (evaluating it multiple times should produce the same
ComputedFacts)

Additionally update test case to cover this use case.