Create a tmp dir for MariaDB environment
We need the ability to mount a tmp dir outside of MariaDB docker
container. We pass it as command line argument to mariadbd instead of a
conf file because it is tied to the path specified in the docker-compose file.
This way the `/tmp/mariadb` location is only specified in one file.

The .env file can now be change to accomodate for the specific
deployment environment.