extractors: Separate server and upload extractors
Because the extractors have different output formats, it is confusing to
expose them under the same interface. Thus, extract_facts is moved from
DataExtractor to ServerFactExtractor and UploadFactExtractor.

Combining different upload facts is different from combining different
server facts, which was not reflected in the old code. This commit
ensures that multiple upload facts for the same upload get combined
correctly.